+++ src/content/focuses.js
@@ -1,7 +1,10 @@
+import * as doms from 'shared/utils/dom';
+
 const focusInput = () => {
   let inputTypes = ['email', 'number', 'search', 'tel', 'text', 'url'];
   let inputSelector = inputTypes.map(type => `input[type=${type}]`).join(',');
-  let target = window.document.querySelector(inputSelector + ',textarea');
+  let targets = window.document.querySelectorAll(inputSelector + ',textarea');
+  let target = Array.from(targets).find(doms.isVisible);
   if (target) {
     target.focus();
   }
